% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{anc_to_density}
\alias{anc_to_density}
\title{Create densities from ancestral}
\usage{
anc_to_density(anc)
}
\arguments{
\item{anc}{an object of class \code{ancestralCauchy}, obtained with \code{\link{ancestral}}.}
}
\value{
list of matrching density objects
}
\description{
Transform an object of class \code{ancestralCauchy}
into a list with \code{density} objects.
}
\keyword{internal}
