% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{ancestral_to_treedata_rec}
\alias{ancestral_to_treedata_rec}
\title{Rectangular posterior density of a node}
\usage{
ancestral_to_treedata_rec(
  tree,
  fit_lat,
  fit_long,
  anc_lat,
  anc_long,
  level = 0.95
)
}
\arguments{
\item{tree}{the phylogenetic tree.}

\item{fit_lat}{a \code{\link{fitCauchy}} object for the latitude trait.}

\item{fit_long}{a \code{\link{fitCauchy}} object for the longitude trait.}

\item{anc_lat}{an object of class \code{ancestralCauchy}, obtained with \code{\link{ancestral}} on \code{fit_lat}.}

\item{anc_long}{an object of class \code{ancestralCauchy}, obtained with \code{\link{ancestral}} on \code{fit_long}.}

\item{level}{the level for the HPD intervals}
}
\value{
An object of class \code{\link[tidytree]{treedata-class}}, that can
be exported with function \code{\link[treeio]{write.beast}}.
}
\description{
Compute the posterior density of a set of node values under a Cauchy process on a phylogenetic tree.
}
\seealso{
\code{\link{ancestral}}, \code{\link{fitCauchy}}
}
\author{
Paul Bastide \email{paul.bastide@m4x.org} and Gilles Didier \email{gilles.didier@free.fr}
}
\keyword{internal}
