% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauphylm.R
\name{cauphylm}
\alias{cauphylm}
\title{Maximum Likelihood estimator for a Cauchy model}
\usage{
cauphylm(
  formula,
  data = list(),
  phy,
  model = c("cauchy", "lambda"),
  lower.bound = list(disp = 0, lambda = 0),
  upper.bound = list(disp = Inf, lambda = NULL),
  starting.value = list(disp = NULL, lambda = NULL),
  hessian = FALSE
)
}
\arguments{
\item{formula}{a model formula.}

\item{data}{a data frame containing variables in the model.
If not found in data, the variables are taken from current environment.}

\item{phy}{a phylogenetic tree of class \code{\link[ape]{phylo}}.}

\item{model}{a model for the trait evolution. Only \code{"cauchy"} is allowed.}

\item{lower.bound}{named list with lower bound values for the parameters.}

\item{upper.bound}{named list with upper bound values for the parameters. See Details.}

\item{starting.value}{starting value for the parameters.
This should be a named list, with \code{disp} the dispersion parameter.
The default initial values are computed from standard statistics used on (independent) Cauchy variables, see details.}

\item{hessian}{if \code{TRUE}, then the numerical hessian is computed, for confidence interval computations. See \code{\link{compute_vcov}}.}
}
\value{
\item{coefficients}{the named vector of coefficients.}
\item{disp}{the maximum likelihood estimate of the dispersion parameter.}
\item{logLik}{the maximum of the log likelihood.}
\item{p}{the number of all parameters of the model.}
\item{aic}{AIC value of the model.}
\item{fitted.values}{fitted values}
\item{residuals}{raw residuals}
\item{y}{response}
\item{X}{design matrix}
\item{n}{number of observations (tips in the tree)}
\item{d}{number of dependent variables}
\item{formula}{the model formula}
\item{call}{the original call to the function}
\item{model}{the phylogenetic model for the covariance}
}
\description{
Find the maximum likelihood, using numerical optimization with \code{\link{optim}}.
}
\details{
Unless specified by the user, the initial values for the parameters are taken according to the following heuristics:
\itemize{
 \item{\code{mu}}{ is the trimmed mean of the trait, keeping only 24\% of the observations, as advocated in Rothenberg et al. 1964 (for models with the intercept only)}
 \item{\code{coef}}{ are obtained from a robust regression using \code{\link[robustbase]{lmrob.S}} (for linear models with several predictors)}
 \item{\code{disp}}{ is initialized from the trait centered and normalized by tip heights, with one of the following statistics, taken from Rousseeuw & Croux 1993:}
 \itemize{
 \item{\code{IQR}}{ half of the inter-quartile range (see \code{\link{IQR}});}
 \item{\code{MAD}}{ median absolute deviation with constant equal to 1 (see \code{\link{mad}});}
 \item{\code{Sn}}{ Sn statistics with constant 0.7071 (see \code{\link[robustbase]{Sn}});}
 \item{\code{Qn}}{ Qn statistics with constant 1.2071 (see \code{\link[robustbase]{Qn}});}
}
}

For \code{model=lambda}, the maximum default value is computed using \code{phytools:::maxLambda},
and is the ratio between the maximum height of a tip node over the maximum height of an internal node.
This can be larger than 1.
}
\references{
Rothenberg T. J., Fisher F. M., Tilanus C. B. 1964. A Note on Estimation from a Cauchy Sample. Journal of the American Statistical Association. 59:460–463.
}
\seealso{
\code{\link{fitCauchy}}, \code{\link[phylolm]{phylolm}}
}
