% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_tree}
\alias{check_tree}
\alias{check_binary_tree}
\title{Check Binary Tree object}
\usage{
check_tree(tree)

check_binary_tree(tree)
}
\arguments{
\item{tree}{a phylogenetic tree}
}
\value{
No value returned.
}
\description{
Perform check on the tree: it needs to be a \code{phylo} object,
with branch lengths, binary, with tip labels.
}
\keyword{internal}
