% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{checkDuplicates}
\alias{checkDuplicates}
\title{Check For Duplicated Entries}
\usage{
checkDuplicates(trait, tree)
}
\arguments{
\item{trait}{(named) vector or matrix of traits being tested.}

\item{tree}{phylogenetic tree.}
}
\description{
Check that the trait are compatible with the tree. Throws an error if not.
Assumes that the trait and tree tips are in the same order, using function 
\code{\link{checkTraitTree}}.
}
\keyword{internal}
