% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{checkTraitTree}
\alias{checkTraitTree}
\title{Check Matrix Parameter}
\usage{
checkTraitTree(trait, tree, name = "trait")
}
\arguments{
\item{trait}{(named) vector or matrix of traits being tested.}

\item{tree}{A phylogenetic tree.}

\item{name}{name of the trait. Default to 'trait'.}
}
\description{
Check that the parameters are compatible with the tree. Throws an error if not.
}
\keyword{internal}
