% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{do_optim}
\alias{do_optim}
\title{NLOPT optimization}
\usage{
do_optim(
  minus_like,
  start.values,
  lower.values,
  upper.values,
  options_nloptr = list(algorithm = "NLOPT_LN_BOBYQA", xtol_rel = 1e-08, maxeval = 1000),
  optim = c("local", "global"),
  ...
)
}
\arguments{
\item{minus_like}{the function to be minimized}

\item{start.values}{vector of starting values}

\item{lower.values}{vector of lower values}

\item{upper.values}{vector of upper values}

\item{options_nloptr}{options to be passed to \code{nloptr}}

\item{optim}{should a global optimization be performed first ?}

\item{...}{further arguments to be passed to minus_like}
}
\description{
Perform the optimization
}
\keyword{internal}
