% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_asr.R
\name{find_closest}
\alias{find_closest}
\title{Find closest value in a vector}
\usage{
find_closest(a, v)
}
\arguments{
\item{a}{a value to be matched}

\item{v}{a vector of grid values}
}
\value{
the index of the closest values of a in v.
}
\description{
Find closest value in a vector
}
\keyword{internal}
