% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cauchy.R
\name{find_longest_tip_branch}
\alias{find_longest_tip_branch}
\title{Find tip with longuest branch}
\usage{
find_longest_tip_branch(tree)
}
\arguments{
\item{tree}{a phylogenetic tree.}
}
\value{
tip with the longuest branch.
}
\description{
Find tip with longuest branch
}
\author{
Paul Bastide \email{paul.bastide@umontpellier.fr} and Gilles Didier \email{gilles.didier@free.fr}
}
\keyword{internal}
