% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_asr.R
\name{find_modes}
\alias{find_modes}
\title{Find modes of a distribution}
\usage{
find_modes(anc, node, mode_proba_method = "integral")
}
\arguments{
\item{anc}{an object of class \code{ancestralCauchy}}

\item{node}{a vector of node number in the tree}
}
\value{
a list, with \code{opt_vals} the abscissa values where the distribution is maximum,
and \code{prop_vals} the approximate density under each optimal value, normalized to sum to one.
}
\description{
Find modes of a distribution
}
\keyword{internal}
