% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCauchy.R
\name{fitCauchyBi}
\alias{fitCauchyBi}
\title{Model fitting for a Cauchy Process}
\usage{
fitCauchyBi(
  phy,
  trait,
  model = c("cauchy", "lambda"),
  method = c("reml", "random.root", "fixed.root"),
  number.params,
  starting.value = list(x0 = NULL, angle = NULL, disp = NULL, lambda = NULL),
  lower.bound = list(angle = NULL, disp = NULL, lambda = NULL),
  upper.bound = list(angle = NULL, disp = NULL, lambda = NULL),
  root.edge = 100,
  hessian = FALSE,
  optim = c("local", "global"),
  method.init.disp = c("Qn", "Sn", "MAD", "IQR")
)
}
\arguments{
\item{phy}{a phylogenetic tree of class \code{\link[ape]{phylo}}.}

\item{trait}{named vector of traits at the tips.}

\item{model}{a model for the trait evolution. One of \code{"cauchy"} or \code{"lambda"} (see Details).}

\item{method}{the method used to fit the process.
One of \code{reml} (the default), \code{fixed.root} or \code{random.root}.
See Details.}

\item{number.params}{TODO}

\item{starting.value}{starting value for the parameters of the Cauchy.
This should be a named list, with \code{x0} and \code{disp} the root starting value and the dispersion parameter.
The default initial values are computed from standard statistics used on (independent) Cauchy variables, see Details.}

\item{lower.bound}{named list with lower bound values for the parameters. See Details for the default values.}

\item{upper.bound}{named list with upper bound values for the parameters. See Details for the default values.}

\item{root.edge}{multiplicative factor for the root dispersion, equal to the length of the root edge. Ignored if \code{method!=random.root}.}

\item{hessian}{if \code{TRUE}, then the numerical hessian is computed, for confidence interval computations. See \code{\link{compute_vcov}}.}

\item{optim}{if "local", only a local optimization around the initial parameter values is performed (the default).
If "global", a global maximization is attempted using the "MLSL" approach (see \code{\link{nloptr}}).}

\item{method.init.disp}{the initialization method for the dispersion.
One of "Qn", "Sn", "MAD", "IQR". Default to the "Qn" statistics.
See Details.}
}
\description{
Fit the bivariate Cauchy process on a phylogeny, using numerical optimization.
}
\references{
Bastide, P. and Didier, G. 2023. The Cauchy Process on Phylogenies: a Tractable Model for Pulsed Evolution. Systematic Biology. doi:10.1093/sysbio/syad053.
}
\seealso{
\code{\link{confint.cauphyfit}}, \code{\link{profile.cauphyfit}}, \code{\link{ancestral}}, \code{\link{increment}}, \code{\link{logDensityTipsCauchy}}, \code{\link{cauphylm}}, \code{\link[geiger]{fitContinuous}}
}
