% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_contour}
\alias{get_contour}
\title{Get Contour}
\usage{
get_contour(den, prob)
}
\arguments{
\item{den}{a two dimensional density object}

\item{prob}{the probaility level}
}
\value{
contour plot coordinates
}
\description{
See \code{\link[emdbook]{HPDregionplot}}
}
\keyword{internal}
