% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_entry_safe}
\alias{get_entry_safe}
\title{Get entry from a list}
\usage{
get_entry_safe(x, i, nn)
}
\arguments{
\item{x}{a list}

\item{i}{the item number}

\item{nn}{the name to retrieve}
}
\value{
entry nn from item i of list x, or NA
}
\description{
Get the entry if the index is acceptable, otherwise return NA.
}
\keyword{internal}
