% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_line_safe}
\alias{get_line_safe}
\title{Get line of a matrix}
\usage{
get_line_safe(x, i)
}
\arguments{
\item{x}{a matrix}

\item{i}{the line number}
}
\value{
the line number i of matrix x, or NA
}
\description{
Get the line if the index is acceptable, otherwise return NA.
}
\keyword{internal}
