% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_asr.R
\name{get_p_value_thermo}
\alias{get_p_value_thermo}
\title{Find the p values of increments}
\usage{
get_p_value_thermo(x, thermo_inc, edge_inc)
}
\arguments{
\item{x}{a cauchy fitted object}

\item{thermo_inc}{a thermo object}

\item{edge_inc}{matching edges}
}
\value{
a new thermo object, with signed pvalues as values.
}
\description{
This function fits a standard BM on the tree, and then use the infered variance
to detect "abnormal" increments.
For each branch i, the "pvalue" of a mode is given by the probability of observing
the absolute value of this mode or higher assuming a Gaussian distribution
with mean 0 and variance sigma_BM * t_i.
The pvalue is signed to the sign of the mode.
}
\keyword{internal}
