% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{getBounds}
\alias{getBounds}
\title{Get bounds}
\usage{
getBounds(model, phy, X, y, values, default.values)
}
\arguments{
\item{model}{the model}

\item{phy}{the phylogenetic tree}

\item{X}{model matrix}

\item{y}{the response vector}

\item{values}{the input values for the bound}

\item{default.values}{the default values for the bound}
}
\description{
Get bounds given the model.
}
\keyword{internal}
