% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{getBoundsCauchy}
\alias{getBoundsCauchy}
\title{Get bound for a Cauchy process}
\usage{
getBoundsCauchy(phy, X, y, values, default.values)
}
\arguments{
\item{phy}{the phylogenetic tree}

\item{X}{model matrix}

\item{y}{the response vector}

\item{values}{the input values for the bound}

\item{default.values}{the default values for the bound}
}
\description{
Get bounds for a Cauchy process
}
\keyword{internal}
