% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{getBoundsLambda}
\alias{getBoundsLambda}
\title{Get bounds for a Cauchy Lambda}
\usage{
getBoundsLambda(phy, X, y, values, default.values)
}
\arguments{
\item{phy}{the phylogenetic tree}

\item{X}{model matrix}

\item{y}{the response vector}

\item{values}{the input values for the bound}

\item{default.values}{the default values for the bound}
}
\description{
Get bounds for a Cauchy Lambda process
}
\keyword{internal}
