% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{getParamNames}
\alias{getParamNames}
\title{Get parameter names}
\usage{
getParamNames(number.params)
}
\arguments{
\item{number.params}{named list with number of parameters.}
}
\description{
Get the names of the parameters depending on the model.
}
\keyword{internal}
