% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{getParamNamesUni}
\alias{getParamNamesUni}
\title{Get parameter names}
\usage{
getParamNamesUni(model, X)
}
\arguments{
\item{model}{model}

\item{X}{model matrix}
}
\description{
Get the names of the parameters depending on the model, for a univariate trait.
}
\keyword{internal}
