% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{getStartingValues}
\alias{getStartingValues}
\title{Get starting values}
\usage{
getStartingValues(model, phy, X, y, starting.value, method.init.disp, method)
}
\arguments{
\item{model}{the model}

\item{phy}{the phylogenetic tree}

\item{X}{model matrix}

\item{y}{the response vector}

\item{starting.value}{the input starting values}
}
\description{
Get starting values given the model.
}
\keyword{internal}
