% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{getStartingValuesLambda}
\alias{getStartingValuesLambda}
\title{Get starting values for a Cauchy Lambda}
\usage{
getStartingValuesLambda(phy, X, y, starting.value, method.init.disp, method)
}
\arguments{
\item{phy}{the phylogenetic tree}

\item{X}{model matrix}

\item{y}{the response vector}

\item{starting.value}{the input starting values}
}
\description{
Get starting values for a Cauchy Lambda process
}
\keyword{internal}
