% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{initDispersionParameter}
\alias{initDispersionParameter}
\title{Initialization of the dispersion parameter.}
\usage{
initDispersionParameter(
  tree,
  trait,
  center,
  method.init.disp = c("Qn", "Sn", "MAD", "IQR"),
  method
)
}
\arguments{
\item{trait}{named vector of traits at the tips.}

\item{center}{the center parameter of the distribution}

\item{method.init.disp}{the robust estimator method}

\item{method}{the method used to fit the process.
One of \code{reml} (the default), \code{fixed.root} or \code{random.root}.
See Details.}
}
\value{
The initial dispersion parameter.
}
\description{
Initialize the dispersion parameter.
}
\details{
Constants are taken from Rousseeuw & Croux, 1993.
}
\references{
Rousseeuw P.J., Croux C. 1993. Alternatives to the Median Absolute Deviation. Journal of the American Statistical Association. 88:1273–1283.
}
\keyword{internal}
