% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{initPositionParameter}
\alias{initPositionParameter}
\title{Initialization of the position parameter.}
\usage{
initPositionParameter(trait)
}
\arguments{
\item{trait}{named vector of traits at the tips.}
}
\value{
The initial position parameter.
}
\description{
Initialize using the trimmed mean of the trait.
}
\references{
Rothenberg T. J., Fisher F. M., Tilanus C. B. 1964. A Note on Estimation from a Cauchy Sample. Journal of the American Statistical Association. 59:460–463.
Rousseeuw P.J., Croux C. 1993. Alternatives to the Median Absolute Deviation. Journal of the American Statistical Association. 88:1273–1283.
}
\keyword{internal}
