% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lizards}
\alias{lizards}
\title{Greater Antillean Anolis lizard dataset}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{phy}{Bayesian maximum clade credibility chronogram for Greater Antillean Anolis from Mahler et al. 2013}
  \item{svl}{Natural log-transformed species average snout-to-vent length}
  \item{ecomorph}{Ecomorph assignments for each of the species}
}
}
\source{
\doi{10.5061/dryad.9g182}
}
\usage{
lizards
}
\description{
A dataset containing the dated phylogeny and the log snout-to-vent length for
Greater Antillean Anolis lizard species,
taken from Mahler et al. 2013.
}
\references{
Mahler, D. Luke; Ingram, Travis; Revell, Liam J.; Losos, Jonathan B. (2013), Data from: Exceptional convergence on the macroevolutionary landscape in island lizard radiations, Dryad, Dataset, https://doi.org/10.5061/dryad.9g182
}
\keyword{datasets}
