% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{logDensityCauchyBi}
\alias{logDensityCauchyBi}
\title{Log Density of a Bivariate Cauchy Process}
\usage{
logDensityCauchyBi(values, center, disp, angle)
}
\arguments{
\item{values}{a matrix with two columns of values where to evaluate the density.}

\item{disp}{vector of the 2 dispersion values along the two axes.}

\item{angle}{vector of the 2 angles of the 2 axes.}

\item{centervector}{of the 2 angles of the 2 axes.}
}
\value{
the log density values for each row of \code{values}.
}
\description{
Compute the log density of the vector of trait at the tips of the phylogenetic tree, 
assuming a Cauchy process.
}
\keyword{internal}
