% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cauchy.R
\name{logDensityTipsCauchy}
\alias{logDensityTipsCauchy}
\title{Log Density of a Cauchy Process}
\usage{
logDensityTipsCauchy(
  tree,
  tipTrait,
  start = NULL,
  disp,
  method = c("reml", "random.root", "fixed.root"),
  rootTip = NULL
)
}
\arguments{
\item{tree}{a phylogeny in \code{\link{ape}} \code{\link[ape]{phylo}} format.}

\item{tipTrait}{a names vector of tip trait values, with names matching the tree labels.}

\item{start}{the initial root trait value.}

\item{disp}{the dispersion parameter of the Cauchy process.}

\item{method}{the method used to fit the process. One of \code{reml} (the default), \code{random.root} or \code{fixed.root}.
See details.}

\item{rootTip}{the tip used to re-root the tree, when the REML method is used. If \code{NULL}, the tip with the longest branch is used. Ignored in `method != "reml"`.}
}
\value{
the density value.
}
\description{
Compute the log density of a Cauchy process on a phylogenetic tree.
}
\examples{
phy <- ape::rphylo(5, 0.1, 0)
dat <- rTraitCauchy(n = 1, phy = phy, model = "cauchy", parameters = list(root.value = 0, disp = 1))
logDensityTipsCauchy(phy, dat, 0, 1, method = "fixed.root")

}
\author{
Paul Bastide \email{paul.bastide@umontpellier.fr} and Gilles Didier \email{gilles.didier@free.fr}
}
