% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{logDensityTipsCauchyBi}
\alias{logDensityTipsCauchyBi}
\title{Log Density of a Bivariate Cauchy Process}
\usage{
logDensityTipsCauchyBi(
  tree,
  tipTrait,
  root.value = NULL,
  disp,
  angle,
  method = c("reml", "random.root", "fixed.root"),
  rootTip = NULL
)
}
\arguments{
\item{tree}{a phylogenetic tree of class \code{\link[ape]{phylo}}.}

\item{tipTrait}{a named matrix of tip trait values, with row names matching the tree labels.}

\item{root.value}{the root starting value of the process.}

\item{disp}{vector of the 2 dispersion values along the two axes.}

\item{angle}{vector of the 2 angles of the 2 axes.}

\item{method}{the method used to compute the likelihood.
One of \code{reml} (the default), \code{fixed.root} or \code{random.root}.}

\item{rootTip}{the tip used to re-root the tree, when the REML method is used.
If \code{NULL}, the tip with the smallest distance to all the others is used (see Details).
Ignored in \code{method != "reml"}.}
}
\value{
the log density value.
}
\description{
Compute the log density of the vector of trait at the tips of the phylogenetic tree, 
assuming a Cauchy process.
}
\seealso{
\code{\link{fitCauchy}}
}
\keyword{internal}
