% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{maxLambda}
\alias{maxLambda}
\title{Maximum lambda value}
\usage{
maxLambda(phy)
}
\arguments{
\item{phy}{the phylogenetic tree}
}
\description{
Find maximum lambda value.
Function taken from \code{phytools:::maxLambda}.
}
\keyword{internal}
