% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{minusLikelihoodFixedRoot}
\alias{minusLikelihoodFixedRoot}
\title{Minus Likelihood function for a Cauchy model}
\usage{
minusLikelihoodFixedRoot(Xdesign)
}
\arguments{
\item{Xdesign}{the design matrix}
}
\value{
A function
}
\description{
Gives the minus likelihood function, with fixed root.
}
\keyword{internal}
