% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{minusLikelihoodFixedRoot_mu}
\alias{minusLikelihoodFixedRoot_mu}
\title{Minus Likelihood function for a Cauchy model}
\usage{
minusLikelihoodFixedRoot_mu(
  param,
  param_names,
  tree,
  trait,
  Xdesign,
  model,
  rootTip
)
}
\arguments{
\item{param}{the parameters where to evaluate the function}

\item{param_names}{the parameter names}

\item{Xdesign}{the design matrix}

\item{model}{a model for the trait evolution. One of \code{"cauchy"} or \code{"lambda"} (see Details).}
}
\value{
The value of minus the log-likelihood.
}
\description{
Gives the minus likelihood function, with fixed root.
}
\keyword{internal}
