% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{minusLikelihoodRandomRoot}
\alias{minusLikelihoodRandomRoot}
\title{Minus Likelihood function for a Cauchy model}
\usage{
minusLikelihoodRandomRoot(pdim)
}
\arguments{
\item{pdim}{dimension of the process}
}
\value{
The value of minus the log-likelihood.
}
\description{
Gives the minus likelihood function, with random root.
}
\keyword{internal}
