% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{minusLikelihoodREML}
\alias{minusLikelihoodREML}
\title{Minus REML function for a Cauchy model}
\usage{
minusLikelihoodREML(pdim)
}
\arguments{
\item{pdim}{dimension of the process}
}
\value{
The value of minus the log-REML.
}
\description{
Gives the minus REML function.
}
\keyword{internal}
