% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{minusLikelihoodREML_uni}
\alias{minusLikelihoodREML_uni}
\title{Minus REML function for a Cauchy model}
\usage{
minusLikelihoodREML_uni(
  param,
  param_names,
  tree,
  trait,
  Xdesign,
  model,
  rootTip
)
}
\arguments{
\item{param}{the parameters where to evaluate the function}

\item{param_names}{the parameter names}

\item{Xdesign}{the design matrix}

\item{model}{a model for the trait evolution. One of \code{"cauchy"} or \code{"lambda"} (see Details).}
}
\value{
The value of minus the log-REML.
}
\description{
Gives the minus REML function.
}
\keyword{internal}
