% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cauchy.R
\name{plot.ancestralCauchy}
\alias{plot.ancestralCauchy}
\title{Plot for class \code{ancestralCauchy}}
\usage{
\method{plot}{ancestralCauchy}(x, node, n.col, ...)
}
\arguments{
\item{x}{an object of class \code{ancestralCauchy}, result of function
\code{\link{ancestral}}.}

\item{node}{the vector of nodes where to plot the ancestral reconstruction.}

\item{n.col}{the number of columns on which to display the plot. Can be left blank.}

\item{...}{further arguments to be passed to \code{\link{plot}}.}
}
\description{
This function takes an object of class \code{ancestralCauchy}, result of function
\code{\link{ancestral}}, and plots the reconstructed states for given nodes.
}
\seealso{
\code{\link{ancestral}}, \code{\link{fitCauchy}}
}
