% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.invisible}
\alias{plot.invisible}
\title{Invisible Plotting}
\usage{
\method{plot}{invisible}(...)
}
\arguments{
\item{...}{arguments to be passed to the plot function.}
}
\value{
the result of the plot, without displaying it.
}
\description{
Invisible Plotting
}
\details{
See https://stackoverflow.com/a/20363489
}
\keyword{internal}
