% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cauchy.R
\name{posteriorDensityAncestral}
\alias{posteriorDensityAncestral}
\title{Posterior density of a node}
\usage{
posteriorDensityAncestral(
  node,
  vals,
  tree,
  tipTrait,
  start = NULL,
  disp,
  method = c("reml", "random.root", "fixed.root")
)
}
\arguments{
\item{node}{the node for which to compute the posterior density.}

\item{vals}{the table of values where the density should be computed.}

\item{tree}{a phylogeny in \code{\link{ape}} \code{\link[ape]{phylo}} format.}

\item{tipTrait}{a names vector of tip trait values, with names matching the tree labels.}

\item{start}{the initial root trait value.}

\item{disp}{the dispersion parameter of the Cauchy process.}

\item{method}{the method used to fit the process. One of \code{reml} (the default), \code{random.root} or \code{fixed.root}.
See details.}
}
\value{
the posterior density value.
}
\description{
Compute the posterior density of a set of node values under a Cauchy process on a phylogenetic tree.
}
\examples{
phy <- ape::rphylo(5, 0.1, 0)
dat <- rTraitCauchy(n = 1, phy = phy, model = "cauchy", parameters = list(root.value = 0, disp = 1))
posteriorDensityAncestral(7, 0.1, phy, dat, disp = 1)

}
\author{
Paul Bastide \email{paul.bastide@umontpellier.fr} and Gilles Didier \email{gilles.didier@free.fr}
}
