% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cauchy.R
\name{printRTreeTest}
\alias{printRTreeTest}
\title{Print a tree}
\usage{
printRTreeTest(tree)
}
\arguments{
\item{tree}{a phylogeny in \code{\link{ape}} \code{\link[ape]{phylo}} format.}
}
\value{
No value returned.
}
\description{
\code{printRtree} prints a tree in Newick format
}
\author{
Paul Bastide \email{paul.bastide@umontpellier.fr} and Gilles Didier \email{gilles.didier@free.fr}
}
\keyword{internal}
\keyword{programming}
