% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reroottip}
\alias{reroottip}
\title{Re root tree at a tip}
\usage{
reroottip(tree, tip)
}
\arguments{
\item{...}{dots arguments to be captured}
}
\value{
a named list of the arguments in ...
}
\description{
Re root tree at a tip, taking care of the root length.
}
\keyword{internal}
