% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_get}
\alias{safe_get}
\title{Safely get element of a named vector}
\usage{
safe_get(x, name)
}
\arguments{
\item{x}{a named vector}

\item{name}{the name of the element to retrieve.}
}
\description{
If the element does not exist, return NULL.
}
\keyword{internal}
