% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateTrait.R
\name{simulateLevyConditional}
\alias{simulateLevyConditional}
\title{Simulate a Levy process}
\usage{
simulateLevyConditional(
  phy,
  brownianVar,
  jumpVar,
  nJumps,
  ancestor,
  root.value
)
}
\arguments{
\item{phy}{an object of class \code{"phylo"}.}

\item{brownianVar}{the variance of the Brownian part of the process.}

\item{jumpVar}{the variance of the Gaussian jumps.}

\item{nJumps}{total number of jumps.}

\item{ancestor}{a logical value specifying whether to return the
    values at the nodes as well (by default, only the values at the tips
    are returned).}

\item{root.value}{a numeric giving the value at the root.}
}
\value{
A named vector of simulated values.
}
\description{
This function uses \code{\link[ape]{rTraitCont}} to simulate a Levy process
with Brownian variance \code{brownianVar}, 
conditionally on \code{nJumps} occurring on the phylogeny,
each jumps Gaussian with variance \code{jumpVar}.
}
\references{
Duchen P., Leuenberger C., Szilágyi S. M., Harmon L. J., Eastman J. M., Schweizer M., Wegmann D. 2017. Inference of Evolutionary Jumps in Large Phylogenies using Lévy Processes. Systematic Biology. 66:950–963.
}
\seealso{
\code{\link[ape]{rTraitCont}}
}
