% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cauchy.R
\name{simulateTipsCauchy}
\alias{simulateTipsCauchy}
\title{Simulate using the Cauchy Process}
\usage{
simulateTipsCauchy(tree, start, disp)
}
\arguments{
\item{tree}{a phylogeny in \code{\link{ape}} \code{\link[ape]{phylo}} format.}

\item{start}{the initial root trait value.}

\item{disp}{the dispersion parameter of the Cauchy process.}
}
\value{
a vector of simulated values.
}
\description{
Simulate tip values with a Cauchy process
}
\author{
Paul Bastide \email{paul.bastide@umontpellier.fr} and Gilles Didier \email{gilles.didier@free.fr}
}
\keyword{internal}
