% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cauchy.R
\name{simulateTipsCauchyOU}
\alias{simulateTipsCauchyOU}
\title{Simulate using the OU Cauchy Process}
\usage{
simulateTipsCauchyOU(tree, start, disp, str, mean)
}
\arguments{
\item{tree}{a phylogeny in \code{\link{ape}} \code{\link[ape]{phylo}} format.}

\item{start}{the initial root trait value.}

\item{disp}{the dispersion parameter of the Cauchy process.}

\item{str}{the selection strength of the OU.}

\item{mean}{the optimal value of the OU.}
}
\value{
a vector of simulated values.
}
\description{
Simulate tip values with a OU Cauchy process
}
\author{
Paul Bastide \email{paul.bastide@umontpellier.fr} and Gilles Didier \email{gilles.didier@free.fr}
}
\keyword{internal}
