% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{transformBranchLengths}
\alias{transformBranchLengths}
\title{Transform branch lengths}
\usage{
transformBranchLengths(phy, model, param)
}
\arguments{
\item{phy}{the phylogenetic tree}

\item{model}{the model}

\item{param}{the parameters}
}
\description{
Transform branch lengths for pagel lambda model
}
\keyword{internal}
