% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{write.evolaps}
\alias{write.evolaps}
\title{Write treedata in evolaps format}
\usage{
write.evolaps(treedata, file)
}
\arguments{
\item{treedata}{an object of class \code{\link[tidytree]{treedata-class}}}

\item{file}{the path of the file}
}
\description{
Write treedata in evolaps format
}
\keyword{internal}
