% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauphylm.R
\name{compute_vcov}
\alias{compute_vcov}
\title{Compute Approximated Variance Covariance Matrix}
\usage{
compute_vcov(obj)
}
\arguments{
\item{obj}{a fitted object, either with \code{\link{cauphylm}} or \code{\link{fitCauchy}}}
}
\value{
The same object, with added vcov entry.
}
\description{
Find the approximated vcov matrix of the parameters.
}
\details{
This function uses the numerical inverse of the Hessian of the likelihood at the optimal value
to approximate the variance covariance matrix.
It can be used to compute confidence intervals with functions \code{\link{confint.cauphylm}}
or \code{\link{confint.cauphyfit}}
}
