% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{getParamNames}
\alias{getParamNames}
\title{Get parameter names}
\usage{
getParamNames(model, X)
}
\arguments{
\item{model}{model}

\item{X}{model matrix}
}
\description{
Get the names of the parameters depending on the model.
}
\keyword{internal}
