% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{minusLikelihoodRandomRoot}
\alias{minusLikelihoodRandomRoot}
\title{Minus Likelihood function for a Cauchy model}
\usage{
minusLikelihoodRandomRoot(
  param,
  param_names,
  tree,
  trait,
  Xdesign,
  model,
  rootTip
)
}
\arguments{
\item{param}{the parameters where to evaluate the function}

\item{param_names}{the parameter names}

\item{Xdesign}{the design matrix}

\item{model}{a model for the trait evolution. One of \code{"cauchy"} or \code{"lambda"} (see Details).}
}
\value{
The value of minus the log-likelihood.
}
\description{
Gives the minus likelihood function, with random root.
}
\keyword{internal}
