% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reroottip}
\alias{reroottip}
\title{Re root tree at a tip}
\usage{
reroottip(tree, tip)
}
\arguments{
\item{tree}{the original tree}

\item{tip}{the tip to re-root at}
}
\value{
the re-rooted tree at the tip
}
\description{
Re root tree at a tip, taking care of the root length.
This function is only used for testing purposes.
}
\keyword{internal}
