% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTraitCauchy.R
\name{simulateTipsCauchy}
\alias{simulateTipsCauchy}
\title{Simulate using the Cauchy Process}
\usage{
simulateTipsCauchy(tree, root.value, disp)
}
\arguments{
\item{tree}{a phylogeny in \code{\link[ape]{ape}} \code{\link[ape]{phylo}} format.}

\item{root.value}{the initial root trait value.}

\item{disp}{the dispersion parameter of the Cauchy process.}
}
\value{
a vector of simulated values.
}
\description{
Simulate tip values with a Cauchy process
}
\keyword{internal}
