% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCauchy.R
\name{vcov.cauphyfit}
\alias{vcov.cauphyfit}
\alias{logLik.cauphyfit}
\alias{print.logLik.cauphyfit}
\alias{AIC.logLik.cauphyfit}
\alias{AIC.cauphyfit}
\alias{confint.cauphyfit}
\alias{coef.cauphyfit}
\title{Generic Methods for cauphyfit}
\usage{
\method{vcov}{cauphyfit}(object, ...)

\method{logLik}{cauphyfit}(object, ...)

\method{print}{logLik.cauphyfit}(x, ...)

\method{AIC}{logLik.cauphyfit}(object, k = 2, ...)

\method{AIC}{cauphyfit}(object, k = 2, ...)

\method{confint}{cauphyfit}(object, parm, level = 0.95, ...)

\method{coef}{cauphyfit}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"phylolm"}.}

\item{...}{further arguments to methods.}

\item{x}{an object of class \code{"phylolm"}.}

\item{k}{numeric, the penalty per parameter to be used; the default \code{k = 2} is the classical AIC.}

\item{parm}{a specification of which parameters are to be given
    confidence intervals, either a vector of numbers or a vector of
    names.  If missing, all parameters are considered.}

\item{level}{the confidence level required.}
}
\description{
Generic Methods for cauphyfit
}
