% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauphylm.R
\name{vcov.cauphylm}
\alias{vcov.cauphylm}
\alias{logLik.cauphylm}
\alias{AIC.logLik.cauphylm}
\alias{AIC.cauphylm}
\alias{predict.cauphylm}
\alias{confint.cauphylm}
\alias{coef.cauphylm}
\title{Generic Methods for S3 class \code{cauphylm}.}
\usage{
\method{vcov}{cauphylm}(object, ...)

\method{logLik}{cauphylm}(object, ...)

\method{AIC}{logLik.cauphylm}(object, k = 2, ...)

\method{AIC}{cauphylm}(object, k = 2, ...)

\method{predict}{cauphylm}(object, newdata = NULL, ...)

\method{confint}{cauphylm}(object, parm, level = 0.95, ...)

\method{coef}{cauphylm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{cauphylm}.}

\item{...}{further arguments to methods.}

\item{k}{numeric, the penalty per parameter to be used; the default \code{k = 2} is the classical AIC.}

\item{newdata}{an optional data frame to provide the predictor values
  at which predictions should be made. If omitted, the fitted values are used. Currently, predictions are made for new species whose placement in the tree is unknown. Only their covariate information is used. The prediction for the trend model is not currently implemented.}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required.}
}
\description{
Generic Methods for S3 class \code{cauphylm}.
}
\seealso{
\code{\link{cauphylm}}

\code{\link[phylolm]{predict.phylolm}}
}
